Fix issue with error on Linux systems
Fix issue #2.

On Linux, an error occurs in `io.Copy` while waiting for `io.EOF`. However, if
the child process is already done it can happen on a Linux system that the
parent process is still trying to read and an `io.EOF` is never reached.

Introduce mitigation logic to treat a path error of `/dev/ptmx` like `io.EOF`.